MFD-ES works in parallel
Restricted communication of geometrical parameters when exchanging
parts of the mesh between processors.

Added function MarkNormalOrientation to detect situation when ghost
faces are oriented opposite to shared face on owner processor.

Corrected MFD-ES to account for orientation at ghost faces and to flip
sign of normal velocity component on ghost faces when it is known to
differ on master face on owner processor.