Solver update
Added multilevel version of crout-ilu solver with maximum product
transversal preordering on each level. The multilevel version is based
on condition number estimation, factorization of rows with large
condition number or small diagonal is postponed for the next level.

Fixes for near-zero values in maximum product transversal algorithm and
negative path-length.

Fixes for output in DrawMatrix, option for text-based output for small
matrices.